Tweaks to the probe settings.
Bit of hack to use the sled docking setup for the rack, but it works for the auto leveling routine (G29) before a print and properly stows the probe, without having to modify the firmware. Autohome (G28) still leaves the probe down, which seems ok because it's followed by a G29 anyway in the standard startup script.